+++ packages/world/src/codegen/systems/MoveSystemLib.sol
@@ -48,6 +48,10 @@ library MoveSystemLib {
     return CallWrapper(self.toResourceId(), address(0)).moveDirections(caller, directions);
   }
 
+  function moveDirectionsPacked(MoveSystemType self, EntityId caller, uint256 packed) internal {
+    return CallWrapper(self.toResourceId(), address(0)).moveDirectionsPacked(caller, packed);
+  }
+
   function move(CallWrapper memory self, EntityId caller, Vec3[] memory newCoords) internal {
     // if the contract calling this function is a root system, it should use `callAsRoot`
     if (address(_world()) == address(this)) revert MoveSystemLib_CallingFromRootSystem();
@@ -71,6 +75,19 @@ library MoveSystemLib {
       : _world().callFrom(self.from, self.systemId, systemCall);
   }
 
+  function moveDirectionsPacked(CallWrapper memory self, EntityId caller, uint256 packed) internal {
+    // if the contract calling this function is a root system, it should use `callAsRoot`
+    if (address(_world()) == address(this)) revert MoveSystemLib_CallingFromRootSystem();
+
+    bytes memory systemCall = abi.encodeCall(
+      _moveDirectionsPacked_EntityId_uint256.moveDirectionsPacked,
+      (caller, packed)
+    );
+    self.from == address(0)
+      ? _world().call(self.systemId, systemCall)
+      : _world().callFrom(self.from, self.systemId, systemCall);
+  }
+
   function move(RootCallWrapper memory self, EntityId caller, Vec3[] memory newCoords) internal {
     bytes memory systemCall = abi.encodeCall(_move_EntityId_Vec3Array.move, (caller, newCoords));
     SystemCall.callWithHooksOrRevert(self.from, self.systemId, systemCall, msg.value);
@@ -80,6 +97,14 @@ library MoveSystemLib {
     bytes memory systemCall = abi.encodeCall(
       _moveDirections_EntityId_DirectionArray.moveDirections,
       (caller, directions)
+    );
+    SystemCall.callWithHooksOrRevert(self.from, self.systemId, systemCall, msg.value);
+  }
+
+  function moveDirectionsPacked(RootCallWrapper memory self, EntityId caller, uint256 packed) internal {
+    bytes memory systemCall = abi.encodeCall(
+      _moveDirectionsPacked_EntityId_uint256.moveDirectionsPacked,
+      (caller, packed)
     );
     SystemCall.callWithHooksOrRevert(self.from, self.systemId, systemCall, msg.value);
   }
@@ -128,6 +153,10 @@ interface _move_EntityId_Vec3Array {
 
 interface _moveDirections_EntityId_DirectionArray {
   function moveDirections(EntityId caller, Direction[] memory directions) external;
+}
+
+interface _moveDirectionsPacked_EntityId_uint256 {
+  function moveDirectionsPacked(EntityId caller, uint256 packed) external;
 }
 
 using MoveSystemLib for MoveSystemType global;