+++ packages/world/ts/exports/internal.ts
@@ -4,3 +4,4 @@ export * from "../trees";
 export * from "../vec3";
 export * from "../entityid";
 export * from "../orientation";
+export * from "../moveUtils";